requirements.txt
# requirements.txt — RPi dependencies for main.py
flask>=3.0
flask-cors>=4.0
paho-mqtt>=1.6
requests>=2.31
PyQt5>=5.15
PyQtWebEngine>=5.15
qrcode>=8.0